extract repo-list logic, create services
