+++ src/main.ts
@@ -1,11 +1,7 @@
 import { OctokitClient } from './service.js';
 import {
   Arguments,
-  IssuesConnection,
-  IssueStatsResult,
   Logger,
-  PullRequestsConnection,
-  PullRequestStatsResult,
   RepositoryStats,
   RepoStatsResult,
   ProcessedPageState,
@@ -14,62 +10,62 @@ import {
   CommandConfig,
 } from './types.js';
 import { StateManager } from './state.js';
-import {
-  appendFileSync,
-  existsSync,
-  readFileSync,
-  unlinkSync,
-  writeFileSync,
-} from 'fs';
-import { resolve } from 'path';
+import { existsSync, unlinkSync, writeFileSync } from 'fs';
 
 import { withRetry, RetryConfig } from './retry.js';
 import {
   generateRepoStatsFileName,
-  convertKbToMb,
-  checkIfHasMigrationIssues,
-  hasLfsTracking,
   formatElapsedTime,
-  resolveOutputPath,
   applyBatchStaggerDelay,
 } from './utils.js';
+import { initCommand, executeCommand } from './init.js';
+import { parseRepoListInput, hasRepoListInput } from './repo-list.js';
+import { isStandaloneRepoListSourceMode } from './repo-stats-source-mode.js';
 import {
-  initializeCsvFile as initializeCsvFileGeneric,
-  appendCsvRow,
-  readCsvFile,
-  REPO_STATS_COLUMNS,
-} from './csv.js';
-import { initCommand, executeCommand, createClientFromOpts } from './init.js';
-import { needsInstallationLookup } from './auth.js';
+  processRepoListSource,
+  validateRepoListSourceOptions,
+} from './repo-list-service.js';
+import { getRepoListForBatch } from './repo-stats-batch.js';
+import { checkForMissingRepos } from './missing-repos-service.js';
 import {
-  NormalizedRepoListEntry,
-  NormalizedRepoList,
-  RepoListOwnerGroup,
-  createRepoListKey,
-  parseRepoListInput,
-  hasRepoListInput,
-  readRepoListInputLines,
-} from './repo-list.js';
+  analyzeRepositoryStats,
+  buildProcessedRepoKeySet,
+  handleRepoProcessingSuccess,
+  initializeCsvFile,
+  mapToRepoStatsResult,
+  processRepositoryByName,
+  writeResultToCsv,
+} from './repo-stats-service.js';
+
+export { initializeCsvFile, mapToRepoStatsResult, writeResultToCsv };
 
 // --- Command configuration ---
 
 const repoStatsConfig: CommandConfig = {
   logPrefix: 'repo-stats',
+  sourceLabel: 'repo-list',
+  supportsInstallationLookup: false,
   summaryLabel: 'PROCESSING',
   generateFileName: generateRepoStatsFileName,
   initializeCsvFile: initializeCsvFile,
+  processSource: processRepoListSource,
   processOrg: processOrgRepoStats,
 };
 
 // --- Public entry point ---
 
 export async function run(opts: Arguments): Promise<string[]> {
-  if (isStandaloneRepoListMode(opts)) {
-    return runRepoListOnly(opts);
+  const isRepoListSourceMode = isStandaloneRepoListSourceMode(opts);
+  if (isRepoListSourceMode) {
+    validateRepoListSourceOptions(opts);
   }
 
   // Build batch-aware config if batch mode is enabled
   const config = { ...repoStatsConfig };
+  if (!isRepoListSourceMode) {
+    config.processSource = undefined;
+  }
+
   if (opts.batchSize != null) {
     const batchIndex = opts.batchIndex ?? 0;
     config.statePrefix = `batch-${batchIndex}`;
@@ -83,400 +79,6 @@ export async function run(opts: Arguments): Promise<string[]> {
   const context = await initCommand(opts, config);
   const result = await executeCommand(context, config);
   return result.outputFiles;
-}
-
-function isStandaloneRepoListMode(opts: Arguments): boolean {
-  const hasRepoList = hasRepoListInput(opts.repoList);
-  const hasOrgList = Array.isArray(opts.orgList) && opts.orgList.length > 0;
-
-  return hasRepoList && !opts.orgName && !hasOrgList;
-}
-
-function generateRepoListStatsFileName(): string {
-  const timestamp = new Date()
-    .toISOString()
-    .replace(/:/g, '')
-    .replace(/\..+/, '');
-  return `repo-list-all_repos-${timestamp}_ts.csv`;
-}
-
-function usesInstallationAuth(opts: Arguments): boolean {
-  return Boolean(
-    opts.appInstallationId || process.env.GITHUB_APP_INSTALLATION_ID,
-  );
-}
-
-function hasTokenAuth(opts: Arguments): boolean {
-  return Boolean(opts.accessToken || process.env.GITHUB_TOKEN);
-}
-
-function validateRepoListAuthSupport(
-  opts: Arguments,
-  normalizedRepoList: NormalizedRepoList,
-) {
-  if (needsInstallationLookup(opts) && !hasTokenAuth(opts)) {
-    throw new Error(
-      'Standalone --repo-list does not currently support GitHub App auto installation lookup. Provide --app-installation-id for a single-owner list or use a GitHub token.',
-    );
-  }
-
-  if (usesInstallationAuth(opts) && normalizedRepoList.summary.ownerCount > 1) {
-    throw new Error(
-      'Standalone --repo-list with GitHub App installation auth currently supports one owner per run. Use a GitHub token or split the list by owner.',
-    );
-  }
-}
-
-async function runRepoListOnly(opts: Arguments): Promise<string[]> {
-  const normalizedRepoList = parseRepoListInput(opts.repoList);
-  if (normalizedRepoList.entries.length === 0) {
-    throw new Error('--repo-list must contain at least one repository entry');
-  }
-  validateRepoListAuthSupport(opts, normalizedRepoList);
-
-  const { logger, client } = await createClientFromOpts(
-    opts,
-    `repo-list-repo-stats-${new Date().toISOString().split('T')[0]}.log`,
-  );
-  logger.info(
-    `Processing ${normalizedRepoList.summary.uniqueEntryCount} repositories ` +
-      `across ${normalizedRepoList.summary.ownerCount} owner groups from --repo-list`,
-  );
-  if (normalizedRepoList.duplicates.length > 0) {
-    logger.warn(
-      `Ignored ${normalizedRepoList.duplicates.length} duplicate repo-list entr${
-        normalizedRepoList.duplicates.length === 1 ? 'y' : 'ies'
-      }`,
-    );
-  }
-
-  const outputDir = opts.outputDir || 'output';
-  const stateManager = new StateManager(outputDir, 'repo-list', logger);
-  const initResult = stateManager.initialize(
-    opts.resumeFromLastSave || false,
-    opts.forceFreshStart || false,
-  );
-  const processedState = initResult.processedState;
-  let fileName = processedState.outputFileName || '';
-
-  if (initResult.resumeFromLastState && fileName) {
-    logger.info(
-      `Resuming from last repo-list state. Using existing file: ${fileName}`,
-    );
-  } else {
-    const baseFileName = opts.outputFileName || generateRepoListStatsFileName();
-    fileName = await resolveOutputPath(opts.outputDir, baseFileName);
-    initializeCsvFile(fileName, logger);
-    processedState.outputFileName = fileName;
-    stateManager.update(processedState, {});
-    logger.info(`Repo-list results will be saved to file: ${fileName}`);
-  }
-
-  const retryConfig: RetryConfig = {
-    maxAttempts: opts.retryMaxAttempts || 3,
-    initialDelayMs: opts.retryInitialDelay || 1000,
-    maxDelayMs: opts.retryMaxDelay || 30000,
-    backoffFactor: opts.retryBackoffFactor || 2,
-    successThreshold: opts.retrySuccessThreshold || 5,
-  };
-  const processingState = {
-    successCount: 0,
-    retryCount: 0,
-    processedCount: 0,
-  };
-  const processedRepoKeys = buildProcessedRepoKeySet(processedState);
-  const startTime = new Date();
-
-  await withRetry(
-    async () => {
-      let processedCount = 0;
-
-      for (const ownerGroup of normalizedRepoList.groupedByOwner.values()) {
-        processedCount += await processRepoListOwnerGroup({
-          ownerGroup,
-          client,
-          logger,
-          opts,
-          processedState,
-          processedRepoKeys,
-          state: processingState,
-          fileName,
-          stateManager,
-        });
-      }
-
-      processedState.completedSuccessfully = true;
-      stateManager.update(processedState, {});
-      logger.info(
-        `Completed repo-list processing. Processed ${processedCount} new repositories. ` +
-          `Total tracked repositories: ${processedRepoKeys.size}. ` +
-          `Elapsed time: ${formatElapsedTime(startTime, new Date())}. ` +
-          `Output saved to: ${fileName}`,
-      );
-    },
-    retryConfig,
-    (state) => {
-      processingState.retryCount++;
-      processingState.successCount = 0;
-      logger.warn(
-        `Retry attempt ${state.attempt}: Failed while processing repo-list. ` +
-          `Last processed repo: ${processedState.lastProcessedRepo}, ` +
-          `Processed repos count: ${processedState.processedRepos.length}, ` +
-          `Error: ${state.error?.message}`,
-      );
-      stateManager.update(processedState, {});
-    },
-  );
-
-  if (opts.autoProcessMissing) {
-    await processMissingRepoListRepositories({
-      opts,
-      normalizedRepoList,
-      client,
-      logger,
-      processedState,
-      processedRepoKeys,
-      retryConfig,
-      fileName,
-      stateManager,
-    });
-  }
-
-  if (opts.cleanState) {
-    stateManager.cleanup();
-  }
-
-  return [fileName];
-}
-
-export function findMissingRepoListEntries({
-  normalizedRepoList,
-  processedFile,
-}: {
-  normalizedRepoList: NormalizedRepoList;
-  processedFile: string;
-}): NormalizedRepoListEntry[] {
-  const records = readCsvFile(processedFile);
-  const processedCsvKeys = new Set<string>();
-
-  (records as Array<{ Org_Name?: string; Repo_Name?: string }>).forEach(
-    (record) => {
-      if (record.Org_Name && record.Repo_Name) {
-        processedCsvKeys.add(
-          createRepoListKey(record.Org_Name, record.Repo_Name),
-        );
-      }
-    },
-  );
-
-  return normalizedRepoList.entries.filter(
-    (entry) => !processedCsvKeys.has(entry.key),
-  );
-}
-
-function groupRepoListEntriesForProcessing(
-  entries: readonly NormalizedRepoListEntry[],
-): RepoListOwnerGroup[] {
-  const groups = new Map<string, NormalizedRepoListEntry[]>();
-
-  for (const entry of entries) {
-    const ownerEntries = groups.get(entry.ownerKey) ?? [];
-    ownerEntries.push(entry);
-    groups.set(entry.ownerKey, ownerEntries);
-  }
-
-  return [...groups.values()].map((entriesForOwner) => ({
-    owner: entriesForOwner[0].owner,
-    ownerKey: entriesForOwner[0].ownerKey,
-    entries: entriesForOwner,
-  }));
-}
-
-async function processMissingRepoListRepositories({
-  opts,
-  normalizedRepoList,
-  client,
-  logger,
-  processedState,
-  processedRepoKeys,
-  retryConfig,
-  fileName,
-  stateManager,
-}: {
-  opts: Arguments;
-  normalizedRepoList: NormalizedRepoList;
-  client: OctokitClient;
-  logger: Logger;
-  processedState: ProcessedPageState;
-  processedRepoKeys: Set<string>;
-  retryConfig: RetryConfig;
-  fileName: string;
-  stateManager: StateManager;
-}): Promise<void> {
-  logger.info('Checking for missing repositories from --repo-list output...');
-  const missingEntries = findMissingRepoListEntries({
-    normalizedRepoList,
-    processedFile: fileName,
-  });
-
-  if (missingEntries.length === 0) {
-    logger.info(
-      'No missing repo-list repositories found. All requested repositories have output rows.',
-    );
-    return;
-  }
-
-  logger.info(
-    `Found ${missingEntries.length} missing repo-list repositories that need to be processed`,
-  );
-  processedState.completedSuccessfully = false;
-  stateManager.update(processedState, {});
-
-  for (const entry of missingEntries) {
-    processedRepoKeys.delete(entry.key);
-  }
-
-  const missingProcessingState = {
-    successCount: 0,
-    retryCount: 0,
-    processedCount: 0,
-  };
-
-  await withRetry(
-    async () => {
-      let processedCount = 0;
-      for (const ownerGroup of groupRepoListEntriesForProcessing(
-        missingEntries,
-      )) {
-        processedCount += await processRepoListOwnerGroup({
-          ownerGroup,
-          client,
-          logger,
-          opts,
-          processedState,
-          processedRepoKeys,
-          state: missingProcessingState,
-          fileName,
-          stateManager,
-        });
-      }
-
-      logger.info(
-        `Completed processing ${processedCount} missing repo-list repositories`,
-      );
-    },
-    retryConfig,
-    (state) => {
-      missingProcessingState.retryCount++;
-      missingProcessingState.successCount = 0;
-      logger.warn(
-        `Retry attempt ${state.attempt}: Failed while processing missing repo-list repositories. ` +
-          `Error: ${state.error?.message}`,
-      );
-      stateManager.update(processedState, {});
-    },
-  );
-
-  processedState.completedSuccessfully = true;
-  stateManager.update(processedState, {});
-}
-
-export function buildProcessedRepoKeySet(
-  processedState: ProcessedPageState,
-): Set<string> {
-  return new Set(
-    processedState.processedRepos.map((repoName) => repoName.toLowerCase()),
-  );
-}
-
-function createProcessedRepoIdentity({
-  owner,
-  repo,
-  mode,
-}: {
-  owner?: string;
-  repo: string;
-  mode: 'org' | 'repo-list';
-}): string {
-  return mode === 'repo-list' && owner ? createRepoListKey(owner, repo) : repo;
-}
-
-function hasProcessedRepo({
-  processedState,
-  processedRepoKeys,
-  identity,
-  mode,
-}: {
-  processedState: ProcessedPageState;
-  processedRepoKeys: Set<string>;
-  identity: string;
-  mode: 'org' | 'repo-list';
-}): boolean {
-  if (processedRepoKeys.has(identity.toLowerCase())) {
-    return true;
-  }
-
-  return mode === 'org' && processedState.processedRepos.includes(identity);
-}
-
-async function processRepoListOwnerGroup({
-  ownerGroup,
-  client,
-  logger,
-  opts,
-  processedState,
-  processedRepoKeys,
-  state,
-  fileName,
-  stateManager,
-}: {
-  ownerGroup: RepoListOwnerGroup;
-  client: OctokitClient;
-  logger: Logger;
-  opts: Arguments;
-  processedState: ProcessedPageState;
-  processedRepoKeys: Set<string>;
-  state: { successCount: number; retryCount: number; processedCount: number };
-  fileName: string;
-  stateManager: StateManager;
-}): Promise<number> {
-  logger.info(
-    `Processing ${ownerGroup.entries.length} repositories for owner ${ownerGroup.owner}`,
-  );
-  let processedCount = 0;
-
-  for (const entry of ownerGroup.entries) {
-    if (processedRepoKeys.has(entry.key)) {
-      logger.debug(`Skipping already processed repository: ${entry.key}`);
-      continue;
-    }
-
-    try {
-      await processRepositoryByName({
-        entry,
-        client,
-        logger,
-        opts,
-        processedState,
-        processedRepoKeys,
-        state,
-        fileName,
-        stateManager,
-        processedCount: ++state.processedCount,
-      });
-      processedCount += 1;
-    } catch (error) {
-      state.successCount = 0;
-      logger.error(
-        `Failed processing repo ${entry.owner}/${entry.repo}: ${
-          error instanceof Error ? error.message : String(error)
-        }`,
-      );
-      throw error;
-    }
-  }
-
-  return processedCount;
 }
 
 // --- Per-org processing (called by shared executeForOrg via config.processOrg) ---
@@ -686,191 +288,6 @@ async function processMissingRepositories({
   }
 }
 
-export function initializeCsvFile(fileName: string, logger: Logger): void {
-  initializeCsvFileGeneric(fileName, REPO_STATS_COLUMNS, logger);
-}
-
-/**
- * Returns the slice of an organization's repository list for the requested
- * batch.
- *
- * By default, fetches the full list of repository names via a lightweight
- * GraphQL query. When `repoListFile` is provided, reads the list from that
- * file instead — useful for parallel matrix runs that would otherwise have
- * every batch re-paginate the same org list and exhaust the installation
- * rate limit. The file format is one entry per line, either bare repo
- * names (combined with `orgName`) or `owner/repo` pairs; blank lines and
- * lines starting with `#` are ignored.
- *
- * Logs the total repository count and batch count so users know how many
- * batches to run.
- */
-export async function getRepoListForBatch({
-  client,
-  orgName,
-  batchSize,
-  batchIndex,
-  pageSize,
-  logger,
-  repoListFile,
-}: {
-  client: Pick<OctokitClient, 'listOrgRepoNames'>;
-  orgName: string;
-  batchSize: number;
-  batchIndex: number;
-  pageSize: number;
-  logger: Logger;
-  repoListFile?: string;
-}): Promise<string[]> {
-  let allRepos: string[];
-
-  if (repoListFile) {
-    logger.info(
-      `Batch mode: reading repository list from file '${repoListFile}' for org '${orgName}' (batch size: ${batchSize}, batch index: ${batchIndex})`,
-    );
-    allRepos = readRepoListFromFile(repoListFile, orgName, logger);
-  } else {
-    logger.info(
-      `Batch mode: fetching repository list for org '${orgName}' (batch size: ${batchSize}, batch index: ${batchIndex})`,
-    );
-    allRepos = [];
-    for await (const repo of client.listOrgRepoNames(orgName, pageSize)) {
-      allRepos.push(`${repo.owner.login}/${repo.name}`);
-    }
-  }
-
-  const totalRepos = allRepos.length;
-  const totalBatches = Math.ceil(totalRepos / batchSize);
-
-  logger.info(
-    `Total repositories: ${totalRepos}, Total batches: ${totalBatches} (batch size: ${batchSize})`,
-  );
-
-  if (totalRepos === 0) {
-    logger.info(
-      `Organization '${orgName}' has no repositories. Nothing to process.`,
-    );
-    return [];
-  }
-
-  if (batchIndex >= totalBatches) {
-    logger.warn(
-      `Batch index ${batchIndex} is out of range (total batches: ${totalBatches}). No repositories to process.`,
-    );
-    return [];
-  }
-
-  const start = batchIndex * batchSize;
-  const end = Math.min(start + batchSize, totalRepos);
-  const batchRepos = allRepos.slice(start, end);
-
-  logger.info(
-    `Batch ${batchIndex} of ${totalBatches}: processing repositories ${start + 1}-${end} of ${totalRepos}`,
-  );
-
-  return batchRepos;
-}
-
-/**
- * Reads a repo list file and normalises every entry to `owner/repo` form.
- * Each non-empty, non-comment line is either a bare repo name (combined
- * with `orgName`) or already in `owner/repo` form. Entries that do not
- * match `orgName` are dropped with a warning so a misplaced list cannot
- * silently process the wrong organization.
- */
-function readRepoListFromFile(
-  filePath: string,
-  orgName: string,
-  logger: Logger,
-): string[] {
-  const resolved = resolve(process.cwd(), filePath);
-  if (!existsSync(resolved)) {
-    throw new Error(`Batch repo list file not found: ${filePath}`);
-  }
-
-  const raw = readFileSync(resolved, 'utf-8');
-  const result: string[] = [];
-  let skipped = 0;
-
-  for (const rawLine of raw.split('\n')) {
-    const line = rawLine.trim();
-    if (line === '' || line.startsWith('#')) continue;
-
-    let owner: string;
-    let repo: string;
-    if (line.includes('/')) {
-      const parts = line.split('/');
-      if (parts.length !== 2 || !parts[0] || !parts[1]) {
-        logger.warn(
-          `Skipping malformed entry in repo list file: '${rawLine}' (expected 'owner/repo' or bare repo name)`,
-        );
-        skipped++;
-        continue;
-      }
-      [owner, repo] = parts;
-    } else {
-      owner = orgName;
-      repo = line;
-    }
-
-    if (owner.toLowerCase() !== orgName.toLowerCase()) {
-      logger.warn(
-        `Skipping entry '${rawLine}' from repo list file: owner '${owner}' does not match --org-name '${orgName}'`,
-      );
-      skipped++;
-      continue;
-    }
-
-    result.push(`${owner}/${repo}`);
-  }
-
-  if (skipped > 0) {
-    logger.warn(
-      `Skipped ${skipped} entr${skipped === 1 ? 'y' : 'ies'} from repo list file '${filePath}'`,
-    );
-  }
-
-  return result;
-}
-
-async function analyzeRepositoryStats({
-  repo,
-  owner,
-  extraPageSize,
-  client,
-  logger,
-}: {
-  repo: RepositoryStats;
-  owner: string;
-  extraPageSize: number;
-  client: OctokitClient;
-  logger: Logger;
-}): Promise<RepoStatsResult> {
-  logger.info(`Analyzing repository: ${owner}/${repo.name}`);
-
-  // Run issue and PR analysis concurrently
-  const [issueStats, prStats] = await Promise.all([
-    analyzeIssues({
-      owner,
-      repo: repo.name,
-      per_page: extraPageSize,
-      issues: repo.issues,
-      client,
-      logger,
-    }),
-    analyzePullRequests({
-      owner,
-      repo: repo.name,
-      per_page: extraPageSize,
-      pullRequests: repo.pullRequests,
-      client,
-      logger,
-    }),
-  ]);
-
-  return mapToRepoStatsResult(repo, issueStats, prStats);
-}
-
 async function* processRepoStats({
   reposIterator,
   client,
@@ -903,114 +320,6 @@ async function* processRepoStats({
 
     yield result;
   }
-}
-
-async function handleRepoProcessingSuccess({
-  result,
-  processedState,
-  state,
-  opts,
-  client,
-  logger,
-  processedCount,
-  stateManager,
-  processedRepoKey,
-}: {
-  result: RepoStatsResult;
-  processedState: ProcessedPageState;
-  state: { successCount: number; retryCount: number };
-  opts: Arguments;
-  client: OctokitClient;
-  logger: Logger;
-  processedCount: number;
-  stateManager: StateManager;
-  processedRepoKey?: string;
-}): Promise<void> {
-  const successThreshold = opts.retrySuccessThreshold || 5;
-
-  // Track successful processing
-  state.successCount++;
-  if (state.successCount >= successThreshold && state.retryCount > 0) {
-    logger.info(
-      `Reset retry count after ${state.successCount} successful operations`,
-    );
-    state.retryCount = 0;
-    state.successCount = 0;
-  }
-
-  stateManager.update(processedState, {
-    repoName: processedRepoKey ?? result.Repo_Name,
-    lastSuccessfulCursor: processedState.currentCursor,
-  });
-
-  // Check rate limits after configured interval
-  if (processedCount % (opts.rateLimitCheckInterval || 10) === 0) {
-    const rateLimitReached = await checkAndHandleRateLimits({
-      client,
-      logger,
-      processedCount,
-    });
-
-    if (rateLimitReached) {
-      throw new Error(
-        'Rate limit reached. Processing will be paused until limits reset.',
-      );
-    }
-  }
-}
-
-async function processRepositoryByName({
-  entry,
-  client,
-  logger,
-  opts,
-  processedState,
-  processedRepoKeys,
-  state,
-  fileName,
-  stateManager,
-  processedCount,
-}: {
-  entry: Pick<NormalizedRepoListEntry, 'owner' | 'repo' | 'key'>;
-  client: OctokitClient;
-  logger: Logger;
-  opts: Arguments;
-  processedState: ProcessedPageState;
-  processedRepoKeys: Set<string>;
-  state: { successCount: number; retryCount: number };
-  fileName: string;
-  stateManager: StateManager;
-  processedCount: number;
-}): Promise<void> {
-  logger.info(`Processing repository: ${entry.owner}/${entry.repo}`);
-
-  const repoStats = await client.getRepoStats(
-    entry.owner,
-    entry.repo,
-    opts.pageSize != null ? Number(opts.pageSize) : 10,
-  );
-
-  const result = await analyzeRepositoryStats({
-    repo: repoStats,
-    owner: entry.owner,
-    extraPageSize: opts.extraPageSize != null ? Number(opts.extraPageSize) : 25,
-    client,
-    logger,
-  });
-
-  await writeResultToCsv(result, fileName, logger);
-  await handleRepoProcessingSuccess({
-    result,
-    processedState,
-    state,
-    opts,
-    client,
-    logger,
-    processedCount,
-    stateManager,
-    processedRepoKey: entry.key,
-  });
-  processedRepoKeys.add(entry.key);
 }
 
 async function processRepositoriesFromFile({
@@ -1064,17 +373,10 @@ async function processRepositoriesFromFile({
 
   for (const entry of repoList) {
     try {
-      const identity = createProcessedRepoIdentity({
-        repo: entry.repo,
-        mode: 'org',
-      });
+      const identity = entry.repo;
       if (
-        hasProcessedRepo({
-          processedState,
-          processedRepoKeys,
-          identity,
-          mode: 'org',
-        })
+        processedRepoKeys.has(identity.toLowerCase()) ||
+        processedState.processedRepos.includes(identity)
       ) {
         logger.debug(`Skipping already processed repository: ${entry.repo}`);
         continue;
@@ -1252,524 +554,4 @@ async function processRepositories({
     successCount: state.successCount,
     retryCount: state.retryCount,
   };
-}
-
-async function checkAndHandleRateLimits({
-  client,
-  logger,
-  processedCount,
-}: {
-  client: OctokitClient;
-  logger: Logger;
-  processedCount: number;
-}): Promise<boolean> {
-  logger.debug(
-    `Checking rate limits after processing ${processedCount} repositories`,
-  );
-  const rateLimits = await client.checkRateLimits();
-
-  if (
-    rateLimits.graphQLRemaining === 0 ||
-    rateLimits.apiRemainingRequest === 0
-  ) {
-    const limitType =
-      rateLimits.graphQLRemaining === 0 ? 'GraphQL' : 'REST API';
-    logger.warn(
-      `${limitType} rate limit reached after processing ${processedCount} repositories`,
-    );
-
-    if (rateLimits.messageType === 'error') {
-      logger.error(`${rateLimits.message}`);
-      throw new Error(
-        `${limitType} rate limit exceeded and maximum retries reached`,
-      );
-    }
-
-    logger.warn(`${rateLimits.message}`);
-    logger.info(`GraphQL remaining: ${rateLimits.graphQLRemaining}`);
-    logger.info(`REST API remaining: ${rateLimits.apiRemainingRequest}`);
-
-    return true; // indicates rate limit was reached
-  } else {
-    logger.info(
-      `GraphQL remaining: ${rateLimits.graphQLRemaining}, REST API remaining: ${rateLimits.apiRemainingRequest}`,
-    );
-  }
-
-  return false; // indicates rate limit was not reached
-}
-
-export async function writeResultToCsv(
-  result: RepoStatsResult,
-  fileName: string,
-  logger: Logger,
-): Promise<void> {
-  try {
-    const formattedResult = {
-      ...result,
-      Is_Empty: result.Is_Empty?.toString().toUpperCase() || 'FALSE',
-      isFork: result.isFork?.toString().toUpperCase() || 'FALSE',
-      isArchived: result.isArchived?.toString().toUpperCase() || 'FALSE',
-      isTemplate: result.isTemplate?.toString().toUpperCase() || 'FALSE',
-      Has_Wiki: result.Has_Wiki?.toString().toUpperCase() || 'FALSE',
-      Has_LFS: result.Has_LFS?.toString().toUpperCase() || 'FALSE',
-      Auto_Merge_Allowed:
-        result.Auto_Merge_Allowed?.toString().toUpperCase() || 'FALSE',
-      Delete_Branch_On_Merge:
-        result.Delete_Branch_On_Merge?.toString().toUpperCase() || 'FALSE',
-      Merge_Commit_Allowed:
-        result.Merge_Commit_Allowed?.toString().toUpperCase() || 'FALSE',
-      Squash_Merge_Allowed:
-        result.Squash_Merge_Allowed?.toString().toUpperCase() || 'FALSE',
-      Rebase_Merge_Allowed:
-        result.Rebase_Merge_Allowed?.toString().toUpperCase() || 'FALSE',
-      Migration_Issue:
-        result.Migration_Issue?.toString().toUpperCase() || 'FALSE',
-    };
-
-    // Create CSV row manually to maintain strict order
-    const values = [
-      formattedResult.Org_Name,
-      formattedResult.Repo_Name,
-      formattedResult.Is_Empty,
-      formattedResult.Last_Push,
-      formattedResult.Last_Update,
-      formattedResult.isFork,
-      formattedResult.isArchived,
-      formattedResult.isTemplate,
-      formattedResult.Visibility,
-      formattedResult.Repo_Size_mb,
-      formattedResult.Record_Count,
-      formattedResult.Collaborator_Count,
-      formattedResult.Protected_Branch_Count,
-      formattedResult.Ruleset_Count,
-      formattedResult.PR_Review_Count,
-      formattedResult.Milestone_Count,
-      formattedResult.Issue_Count,
-      formattedResult.PR_Count,
-      formattedResult.PR_Review_Comment_Count,
-      formattedResult.Commit_Comment_Count,
-      formattedResult.Issue_Comment_Count,
-      formattedResult.Issue_Event_Count,
-      formattedResult.Release_Count,
-      formattedResult.Project_Count,
-      formattedResult.Branch_Count,
-      formattedResult.Tag_Count,
-      formattedResult.Discussion_Count,
-      formattedResult.Star_Count,
-      formattedResult.Fork_Count,
-      formattedResult.Watcher_Count,
-      formattedResult.Has_Wiki,
-      formattedResult.Has_LFS,
-      formattedResult.Default_Branch,
-      formattedResult.Primary_Language,
-      formattedResult.Languages,
-      formattedResult.License,
-      formattedResult.Topics,
-      formattedResult.Description,
-      formattedResult.Homepage_URL,
-      formattedResult.Auto_Merge_Allowed,
-      formattedResult.Delete_Branch_On_Merge,
-      formattedResult.Merge_Commit_Allowed,
-      formattedResult.Squash_Merge_Allowed,
-      formattedResult.Rebase_Merge_Allowed,
-      formattedResult.Full_URL,
-      formattedResult.Migration_Issue,
-      formattedResult.Created,
-    ];
-
-    appendCsvRow(fileName, values, logger);
-
-    logger.info(
-      `Successfully wrote result for repository: ${result.Repo_Name}`,
-    );
-  } catch (error) {
-    logger.error(
-      `Failed to write CSV for repository ${result.Repo_Name}: ${
-        error instanceof Error ? error.message : String(error)
-      }`,
-    );
-    throw error;
-  }
-}
-
-export function mapToRepoStatsResult(
-  repo: RepositoryStats,
-  issueStats: IssueStatsResult,
-  prStats: PullRequestStatsResult,
-): RepoStatsResult {
-  const repoSizeMb = convertKbToMb(repo.diskUsage);
-  const totalRecordCount = calculateRecordCount(repo, issueStats, prStats);
-  const hasMigrationIssues = checkIfHasMigrationIssues({
-    repoSizeMb,
-    totalRecordCount,
-  });
-
-  // Format languages as a semicolon-separated list with percentages
-  const languagesStr =
-    repo.languages?.edges
-      ?.map((edge) => {
-        const pct =
-          repo.languages.totalSize > 0
-            ? ((edge.size / repo.languages.totalSize) * 100).toFixed(1)
-            : '0.0';
-        return `${edge.node.name}:${pct}%`;
-      })
-      .join(';') ?? '';
-
-  // Format topics as a semicolon-separated list
-  const topicsStr =
-    repo.repositoryTopics?.nodes?.map((t) => t.topic.name).join(';') ?? '';
-
-  return {
-    Org_Name: repo.owner.login.toLowerCase(),
-    Repo_Name: repo.name.toLowerCase(),
-    Is_Empty: repo.isEmpty,
-    Last_Push: repo.pushedAt,
-    Last_Update: repo.updatedAt,
-    isFork: repo.isFork,
-    isArchived: repo.isArchived,
-    isTemplate: repo.isTemplate,
-    Visibility: repo.visibility ?? '',
-    Repo_Size_mb: repoSizeMb,
-    Record_Count: totalRecordCount,
-    Collaborator_Count: repo.collaborators.totalCount,
-    Protected_Branch_Count: repo.branchProtectionRules.totalCount,
-    Ruleset_Count: repo.rulesets.totalCount,
-    PR_Review_Count: prStats.prReviewCount,
-    PR_Review_Comment_Count: prStats.prReviewCommentCount,
-    Commit_Comment_Count: repo.commitComments.totalCount,
-    Milestone_Count: repo.milestones.totalCount,
-    PR_Count: repo.pullRequests.totalCount,
-    Project_Count: repo.projectsV2.totalCount,
-    Branch_Count: repo.branches.totalCount,
-    Release_Count: repo.releases.totalCount,
-    Issue_Count: issueStats.totalIssuesCount,
-    Issue_Event_Count: issueStats.issueEventCount + prStats.issueEventCount,
-    Issue_Comment_Count:
-      issueStats.issueCommentCount + prStats.issueCommentCount,
-    Tag_Count: repo.tags.totalCount,
-    Discussion_Count: repo.discussions.totalCount,
-    Star_Count: repo.stargazerCount ?? 0,
-    Fork_Count: repo.forkCount ?? 0,
-    Watcher_Count: repo.watchers?.totalCount ?? 0,
-    Has_Wiki: repo.hasWikiEnabled,
-    Has_LFS: hasLfsTracking(repo.gitattributes?.text),
-    Default_Branch: repo.defaultBranchRef?.name ?? '',
-    Primary_Language: repo.primaryLanguage?.name ?? '',
-    Languages: languagesStr,
-    License: repo.licenseInfo?.spdxId || repo.licenseInfo?.name || '',
-    Topics: topicsStr,
-    Description: repo.description ?? '',
-    Homepage_URL: repo.homepageUrl ?? '',
-    Auto_Merge_Allowed: repo.autoMergeAllowed ?? false,
-    Delete_Branch_On_Merge: repo.deleteBranchOnMerge ?? false,
-    Merge_Commit_Allowed: repo.mergeCommitAllowed ?? false,
-    Squash_Merge_Allowed: repo.squashMergeAllowed ?? false,
-    Rebase_Merge_Allowed: repo.rebaseMergeAllowed ?? false,
-    Full_URL: repo.url,
-    Migration_Issue: hasMigrationIssues,
-    Created: repo.createdAt,
-  };
-}
-
-function calculateRecordCount(
-  repo: RepositoryStats,
-  issueStats: IssueStatsResult,
-  prStats: PullRequestStatsResult,
-): number {
-  // Match exactly how the bash script calculates record count (line 918)
-  return (
-    repo.collaborators.totalCount +
-    repo.branchProtectionRules.totalCount +
-    prStats.prReviewCount +
-    repo.milestones.totalCount +
-    issueStats.totalIssuesCount +
-    repo.pullRequests.totalCount +
-    prStats.prReviewCommentCount +
-    repo.commitComments.totalCount +
-    issueStats.issueCommentCount +
-    prStats.issueCommentCount +
-    issueStats.issueEventCount +
-    prStats.issueEventCount +
-    repo.releases.totalCount +
-    repo.projectsV2.totalCount
-  );
-}
-
-async function analyzeIssues({
-  owner,
-  repo,
-  per_page,
-  issues,
-  client,
-  logger,
-}: {
-  owner: string;
-  repo: string;
-  per_page: number;
-  issues: IssuesConnection;
-  client: OctokitClient;
-  logger: Logger;
-}): Promise<IssueStatsResult> {
-  logger.debug(`Analyzing issues for repository: ${repo}`);
-
-  if (issues.totalCount <= 0) {
-    logger.debug(`No issues found for repository: ${repo}`);
-    return {
-      totalIssuesCount: issues.totalCount,
-      issueEventCount: 0,
-      issueCommentCount: 0,
-    };
-  }
-
-  let totalEventCount = 0;
-  let totalCommentCount = 0;
-
-  // Process first page
-  for (const issue of issues.nodes) {
-    const eventCount = issue.timeline.totalCount;
-    const commentCount = issue.comments.totalCount;
-
-    // Calculate non-comment events by subtracting comments from total timeline events
-    totalEventCount += eventCount - commentCount;
-    totalCommentCount += commentCount;
-  }
-
-  // Process additional pages if they exist
-  if (issues.pageInfo.hasNextPage && issues.pageInfo.endCursor != null) {
-    logger.debug(`More pages of issues found for repository: ${repo}`);
-
-    try {
-      // Get next page of issues using iterator
-      const nextPagesIterator = client.getRepoIssues(
-        owner,
-        repo,
-        per_page,
-        issues.pageInfo.endCursor,
-      );
-
-      // Process each issue from subsequent pages
-      for await (const issue of nextPagesIterator) {
-        const eventCount = issue.timeline.totalCount;
-        const commentCount = issue.comments.totalCount;
-
-        // Calculate non-comment events by subtracting comments from total timeline events
-        totalEventCount += eventCount - commentCount;
-        totalCommentCount += commentCount;
-      }
-    } catch (error) {
-      logger.error(
-        `Error retrieving additional issues for ${owner}/${repo}. ` +
-          `Consider reducing page size. Error: ${error}`,
-        error,
-      );
-      throw error;
-    }
-  }
-
-  logger.debug(`Gathered all issues from repository: ${repo}`);
-  return {
-    totalIssuesCount: issues.totalCount,
-    issueEventCount: totalEventCount,
-    issueCommentCount: totalCommentCount,
-  };
-}
-
-async function analyzePullRequests({
-  owner,
-  repo,
-  per_page,
-  pullRequests,
-  client,
-  logger,
-}: {
-  owner: string;
-  repo: string;
-  per_page: number;
-  pullRequests: PullRequestsConnection;
-  client: OctokitClient;
-  logger: Logger;
-}): Promise<PullRequestStatsResult> {
-  if (pullRequests.totalCount <= 0) {
-    return {
-      prReviewCommentCount: 0,
-      commitCommentCount: 0,
-      issueEventCount: 0,
-      issueCommentCount: 0,
-      prReviewCount: 0,
-    };
-  }
-
-  let issueEventCount = 0;
-  let issueCommentCount = 0;
-  let prReviewCount = 0;
-  let prReviewCommentCount = 0;
-  let commitCommentCount = 0;
-
-  // Process first page
-  for (const pr of pullRequests.nodes) {
-    const eventCount = pr.timeline.totalCount;
-    const commentCount = pr.comments.totalCount;
-    const reviewCount = pr.reviews.totalCount;
-    const commitCount = pr.commits.totalCount;
-
-    // This matches how the bash script handles event counts
-    // It subtracts comments from timeline events, and handles commit limits
-    const redundantEventCount =
-      commentCount + (commitCount > 250 ? 250 : commitCount);
-
-    const adjustedEventCount = Math.max(0, eventCount - redundantEventCount);
-
-    issueEventCount += adjustedEventCount;
-    issueCommentCount += commentCount;
-    prReviewCount += reviewCount;
-
-    // Count review comments by examining each review
-    for (const review of pr.reviews.nodes) {
-      prReviewCommentCount += review.comments.totalCount;
-    }
-
-    commitCommentCount += commitCount;
-  }
-
-  // Process additional pages if they exist
-  if (
-    pullRequests.pageInfo.hasNextPage &&
-    pullRequests.pageInfo.endCursor != null
-  ) {
-    const cursor = pullRequests.pageInfo.endCursor;
-    logger.debug(
-      `Fetching additional pull requests for ${repo} starting from cursor ${cursor}`,
-    );
-
-    for await (const pr of client.getRepoPullRequests(
-      owner,
-      repo,
-      per_page,
-      cursor,
-    )) {
-      const eventCount = pr.timeline.totalCount;
-      const commentCount = pr.comments.totalCount;
-      const reviewCount = pr.reviews.totalCount;
-      const commitCount = pr.commits.totalCount;
-
-      const redundantEventCount =
-        commentCount + (commitCount > 250 ? 250 : commitCount);
-
-      const adjustedEventCount = Math.max(0, eventCount - redundantEventCount);
-
-      issueEventCount += adjustedEventCount;
-      issueCommentCount += commentCount;
-      prReviewCount += reviewCount;
-
-      // Process review comments for additional pages
-      for (const review of pr.reviews.nodes) {
-        prReviewCommentCount += review.comments.totalCount;
-      }
-
-      commitCommentCount += commitCount;
-    }
-  }
-
-  return {
-    prReviewCommentCount,
-    commitCommentCount,
-    issueEventCount,
-    issueCommentCount,
-    prReviewCount,
-  };
-}
-
-export async function checkForMissingRepos({
-  opts,
-  processedFile,
-}: {
-  opts: Arguments;
-  processedFile: string;
-}): Promise<{
-  missingRepos: string[];
-}> {
-  const logFileName = `${opts.orgName!}-missing-repos-check-${
-    new Date().toISOString().split('T')[0]
-  }.log`;
-  const { logger, client } = await createClientFromOpts(opts, logFileName);
-
-  const org = opts.orgName!.toLowerCase();
-  const per_page = opts.pageSize || 10;
-
-  logger.debug(`Checking for missing repositories in organization: ${org}`);
-
-  logger.info(
-    `Reading processed file: ${processedFile} to check for missing repositories`,
-  );
-  const records = readCsvFile(processedFile);
-
-  logger.debug(`Parsed ${records.length} records from processed file`);
-  const processedReposSet = new Set<string>();
-  (records as Array<{ Repo_Name: string }>).forEach((record) => {
-    processedReposSet.add(record.Repo_Name.toLowerCase());
-  });
-
-  // file name of output file with missing repos with datetime suffix
-  function generateTimestampSuffix(date: Date): string {
-    const iso = date.toISOString();
-    const [datePart, timePart] = iso.split('T');
-    const [hour, minute] = timePart.split(':');
-    return `${datePart}-${hour}-${minute}`;
-  }
-  const timestampSuffix = generateTimestampSuffix(new Date());
-  const baseMissingReposFileName = `${org}-missing-repos-${timestampSuffix}.csv`;
-  const missingReposFileName = await resolveOutputPath(
-    opts.outputDir,
-    baseMissingReposFileName,
-  );
-
-  logger.info('Checking for missing repositories');
-  const missingRepos = [];
-
-  if (hasRepoListInput(opts.repoList)) {
-    // Check missing repos from the provided repo list
-    logger.info('Checking against provided repo list');
-    const repoListRaw = readRepoListInputLines(opts.repoList);
-
-    const repoList = repoListRaw
-      .filter((line) => line.trim() !== '' && !line.trim().startsWith('#'))
-      .map((line) => {
-        const parts = line.trim().split('/');
-        return {
-          owner: parts.length > 1 ? parts[0] : '',
-          repo: parts.length > 1 ? parts[1] : parts[0],
-        };
-      })
-      .filter(({ owner }) => !owner || owner.toLowerCase() === org);
-
-    logger.info(`Found ${repoList.length} repos for ${org} in repo list`);
-
-    for (const { repo: repoName } of repoList) {
-      if (!processedReposSet.has(repoName.toLowerCase())) {
-        missingRepos.push(repoName);
-        const csvRow = `${repoName}\n`;
-        appendFileSync(missingReposFileName, csvRow);
-      }
-    }
-  } else {
-    // Check missing repos from all org repos
-    logger.info('Checking against all organization repositories');
-    for await (const repo of client.listReposForOrg(org, per_page)) {
-      if (processedReposSet.has(repo.name.toLowerCase())) {
-        continue;
-      } else {
-        missingRepos.push(repo.name);
-        const csvRow = `${repo.name}\n`;
-        appendFileSync(missingReposFileName, csvRow);
-      }
-    }
-  }
-  logger.info(`Found ${missingRepos.length} missing repositories`);
-  if (missingRepos.length > 0) {
-    logger.info(`Missing repositories written to: ${missingReposFileName}`);
-  }
-
-  return { missingRepos };
 }
